Move to JDK11, remove all action item from sonaLint static analysis

+++ src/test/java/it/jugtorino/chalk/AmphoraTest.java
@@ -225,7 +225,7 @@ public class AmphoraTest {
 		assertThat(amphora.valueOf("name"), equalTo("2"));
 		assertThat((Integer) amphora.rawValueOf("name"), equalTo(2));
 		assertEquals("2", amphora.valueOf("name"));
-		assertEquals(2, amphora.rawValueOf("name"));
+		assertEquals(2, (int)amphora.rawValueOf("name"));
 
 	}
